fix: disable initialization when toggling hidden RAM settings
This closes #17, though the bug is fixed partially. Apparently the commented `HiddenRAM_MatAryInit()` triggers MCS crash when the affected calculator opens C.Basic and the `Use Hidn RAM` setting in the saved config file is off. What I don't understand is that this function is used here the same way it is used in `main()`, but that one doesn't cause the crash. My thought at the moment is that such initialization at main function takes care of different value of bit array variable `UseHiddenRAM`, while these ones don't.

The issue will be closed for now, unless some new discoveries are made about the actual cause.