feature: abort when toggling hidden RAM settings (experimental)
Related to #17. This is from the fact that `HiddenRAM_MatAryInit()` works well in both on and off cases right after opening C.Basic, but not in the settings page. As such, abort function is required here to let us quit and restart C.Basic with the modified hidden RAM setting.